✨ (gitignore): Add Webjs cache & auth directories to .gitignore
Add the directories .wwebjs_auth/ and .wwebjs_cache/ to the .gitignore file to prevent them from being tracked in the repository.

➕ (rome.json): Add rome.json configuration file

Add a new file, rome.json, with the initial configuration for the Rome tool. This file includes settings for organizing imports, enabling the linter, setting JavaScript formatting options, and configuring the formatter.